test: creates common function for asserting expected moves
